app/requirements.txt
alembic==0.9.9
aniso8601==3.0.2
appnope==0.1.0
Babel==2.5.3
backcall==0.1.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
decorator==4.3.0
Flask==0.12.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
ipython==6.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mimerender==0.6.0
parso==0.2.0
passlib==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.7.2
python-editor==1.0.3
python-mimeparse==1.6.0
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.6
traitlets==4.3.2
uuid==1.30
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.1
